Make runner compatible with rspec3
RSpec 3 deprecated the line option and nested format.  This is
a stepping stone towards getting this working with rspec 3.

This is a little messy because I noticed that the runner itself is
setting defaults (like the rspec binary name and format) that probably
should be entirely up to the configuration object.  Will fix in a
followup